add example of using interface as type
on the main thread entrypoint it swaps from EthereumWallet to SolanaWallet, this is an example of sending for example USDC token.

so the benefit, if the SolanaWallet or EthereumWallet one of the goes down, we can easily change the implementation and bind the new payment method to the variable, without need to refactoring all the shits.